pyproject.toml
[project]
name = "monetrix"
version = "0.1.0"
description = ""
authors = [
    {name = "Aaliyah Harper",email = "aeverly14@pm.me"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
]

[tool.poetry]
packages = [{include = "monetrix", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
pyright = "^1.1.399"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C90", "ANN"]
